Функция для вывода игровых экранов, больше модулей
